correction-tp-test-1/src/app/components/product-list/product-list.component.ts
import { CommonModule } from '@angular/common';
import { Component, Input, Output, EventEmitter } from '@angular/core';

interface Product {
  id: number;
  name: string;
  price: number;
  stock: number;
}

@Component({
  selector: 'app-product-list',
  standalone:true,
  imports:[CommonModule],
  templateUrl: './product-list.component.html',
  styleUrls: ['./product-list.component.css']
})
export class ProductListComponent {
  @Input() products: Product[] = [];
  @Output() editProduct = new EventEmitter<Product>();
  @Output() deleteProduct = new EventEmitter<number>();

  onEdit(product: Product) {
    this.editProduct.emit(product);
  }

  onDelete(productId: number) {
    this.deleteProduct.emit(productId);
  }
}